Split AdminToolsController into per-entity admin controllers
Resolves #8. The single AdminToolsController is split into one controller
per entity grouping (Enemies, Items, ItemMods, Skills, Tags, Zones,
Challenges) under Game.Api/Controllers/Admin, mirroring the frontend
Workbench's entity-driven layout.

Each controller keeps an explicit [Route("/api/AdminTools/[action]")]
template instead of the [controller] token, so the public route surface,
the generated TypeScript client, and the frontend's AdminTools/* endpoint
keys are unchanged. Per-controller dependencies are trimmed to only what
each grouping needs, and every controller retains the [ApiController]
attribute and AdminCacheInvalidationFilter.

Updated the Admin Tools API surface section of docs/backend.md to describe
the new layout and the rationale for the shared route prefix.
