+++ docs/backend.md
@@ -37,13 +37,15 @@ As mentioned above, most of the reference data is cached in-memory on the backen
 
 ## Admin Tools API surface
 
-The frontend admin tools were consolidated into a single entity-driven Workbench (see `docs/frontend.md`). That UI saves a whole record (identity + related collections) at once, but the backend keeps persisting each relationship through its own `AdminToolsController` endpoint — the Workbench just orchestrates them.
+The frontend admin tools were consolidated into a single entity-driven Workbench (see `docs/frontend.md`). That UI saves a whole record (identity + related collections) at once, but the backend keeps persisting each relationship through its own admin endpoint — the Workbench just orchestrates them.
+
+These endpoints are split across one controller per entity grouping (`AdminEnemiesController`, `AdminItemsController`, `AdminItemModsController`, `AdminSkillsController`, `AdminTagsController`, `AdminZonesController`, `AdminChallengesController`, all under `Game.Api/Controllers/Admin`), mirroring the Workbench's entity-driven layout. They deliberately **share the `/api/AdminTools/[action]` route prefix** via an explicit `[Route]` template (rather than the `[controller]` token) so the generated TypeScript client and the frontend's `AdminTools/*` endpoint keys stay stable — the split is purely a backend code-organization concern and does not change the public API surface. Each controller injects only the repositories it needs and still carries the `[ApiController]` + `AdminCacheInvalidationFilter` attributes.
 
 Two persistence gotchas to respect when adding or editing these endpoints:
 
 - **Updates must not be inferred from the key value.** `IEntityStore.Update` tracks the passed entity as `Modified` directly (`_context.Entry(entity).State = EntityState.Modified`) instead of going through `DbContext.Update`. `DbContext.Update` infers Add-vs-Edit from whether the key is "set", which misfires for our zero-based-identity tables (see [Reference Data](#reference-data)): the first row's `Id == 0` is the CLR/identity-column default, so EF would treat an edit of record 0 as a brand-new insert and silently duplicate it.
 - **Don't update/delete cached entities directly.** Reference data is read with `AsNoTracking().Include(...)`, so a fetched child (e.g. a `SkillDamageMultiplier`) carries a loaded back-reference to its parent and siblings. Passing that instance to `Update`/`Delete` would drag the whole graph into the change tracker and re-insert it. Relationship endpoints therefore build fresh, navigation-free entities keyed only by their own ids (as `SetEnemySkills`/`SetZoneEnemies`/`SetSkillMultipliers` do) rather than mutating the cached objects.
-- **Invalidate the read-cache after a write.** Each static-reference repository caches its list and exposes `InvalidateCache()`; `AdminCacheInvalidationFilter` (a controller-level filter on `AdminToolsController`) clears every such cache after a successful action so the next read returns fresh data. A new admin entity must add its repository to that filter. `AddEditChallenges` follows this exactly: a `Challenge` carries no child relationships, so it has a single whole-record Add/Edit/Delete endpoint (no `Set*` endpoints) and the only follow-up is the cache invalidation. Its statistic/entity are derived from the `ChallengeType` and never persisted, so the endpoint binds and stores only `ChallengeTypeId` + the goal/target/reward fields.
+- **Invalidate the read-cache after a write.** Each static-reference repository caches its list and exposes `InvalidateCache()`; `AdminCacheInvalidationFilter` (a controller-level filter applied to every admin controller under `Game.Api/Controllers/Admin`) clears every such cache after a successful action so the next read returns fresh data. A new admin entity must add its repository to that filter. `AddEditChallenges` follows this exactly: a `Challenge` carries no child relationships, so it has a single whole-record Add/Edit/Delete endpoint (no `Set*` endpoints) and the only follow-up is the cache invalidation. Its statistic/entity are derived from the `ChallengeType` and never persisted, so the endpoint binds and stores only `ChallengeTypeId` + the goal/target/reward fields.
 
 ## Integration-Test Containers in Constrained Environments
